Compatible with python2

+++ models/densenet.py
@@ -1,3 +1,4 @@
+from __future__ import division
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
@@ -143,8 +144,8 @@ class DenseNet(nn.Module):
     def forward(self, x):
         features = self.features(x)
         out = F.relu(features, inplace=True)
-        last_duration = math.ceil(self.sample_duration / 16)
-        last_size = math.floor(self.sample_size / 32)
+        last_duration = int(math.ceil(self.sample_duration / 16))
+        last_size = int(math.floor(self.sample_size / 32))
         out = F.avg_pool3d(out, kernel_size=(last_duration, last_size, last_size)).view(features.size(0), -1)
         if self.last_fc:
             out = self.classifier(out)
